More stats
Count how many of the executions were requested by text message
(and how many of those were two-parts), inline query and sending documents.

Also count executions in which stdin is present.

I'm adding this to be able to have a better understanding
of how the bot is used.

I'm considering redesigning inline mode in next release, and that data
will probably help me decide if and how to do it.