Modified langs.json to work better with /alias
Made python, lisp and c++ unambiguous,
let users know they can adjust their meaning.

"Hardlinked" default aliases which should probably always mean the same
(like c++ and cplusplus, and js and javascript).

It's not possible for users to separate hardlinked aliases,
and I think it won't be. I'm more likely to separate them
if they turn out to not be as synonymous as I thought.