[Bugfix] bot only sends typing before actually replying
The bot was sending 'typing' when seeing *any* command,
even if it wasn't recognized as language name,
and bot was not gonna reply to it.

execute now returns event emitter,
which can be converted back to promise
using constants.emitter-to-promise.

The event emitter emits 'language-resolved' if language gets resolved
and code execution is about to be started.

The change caused responder.preparing-response-to
to be called AFTER responder.respond-object,
responder.preparing-response-to had to be altered.

Code became slightly more messy after the fix,
it's time to think about rewriting that part of the code
(responder, objects.respond).

The bug was introduced when adding Redis and customization to the bot
-- before Redis, language resolving was synchronous, and if it failed,
execution promise was synchronously rejected;
'typing' was not send if execution.is-pending! retuned false.
After adding Redis customization, resolving language became async,
which made synchronous reject impossible, which made the check
always return true, which caused the bot to always send 'typing'.