Display null as symbol for null, replace « with <<
Trying to send null character causes the Telegram Bot API
to error out with cryptic message.

And Telegram Desktop replaces << with «
and >> with » for some reason, when not in monospace,
and that causes errors while trying to use bitshift operators.
So, I replace them back.
I cannot think of anyone *wanting* to use « or ».